Server_FASTAPI_Unused/requirements.txt
fastapi[all]~=0.103.2
psycopg2-binary
python-multipart
python-jose[cryptography]
passlib[bcrypt]~=1.7.4
sqlalchemy~=2.0.21
uvicorn~=0.23.2
pydantic~=2.4.2

